+++ resources/templates/blade/complete/404.blade.php
@@ -1,0 +1,7 @@
+@extends('layouts.default')
+
+@section('wrapper')
+  <div class="wrapper text-center d-flex flex-column justify-content-center" style="height:100vh;">
+    <h1>404</h1> 
+  </div>
+@endsection