+++ resources/templates/php/complete/header.php
@@ -27,7 +27,6 @@ $container = get_theme_mod( 'understrap_container_type' );
 
 <div class="hfeed site" id="page">
 
-	<!-- ******************* The Navbar Area ******************* -->
 	<div class="wrapper-fluid wrapper-navbar" id="wrapper-navbar">
 
 		<a class="skip-link screen-reader-text sr-only" href="#content"><?php esc_html_e( 'Skip to content', 'understrap' ) ?></a>
@@ -42,7 +41,6 @@ $container = get_theme_mod( 'understrap_container_type' );
 					<span class="navbar-toggler-icon"></span>
 				</button>
 
-					<!-- Your site title as branding in the menu -->
 					<?php if ( ! has_custom_logo() ) { ?>
 
 						<?php if ( is_front_page() && is_home() ) : ?>
@@ -58,9 +56,8 @@ $container = get_theme_mod( 'understrap_container_type' );
 					
 					<?php } else {
 						the_custom_logo();
-					} ?><!-- end custom logo -->
+					} ?>
 
-				<!-- The WordPress Menu goes here -->
 				<?php wp_nav_menu(
 					array(
 						'theme_location'  => 'primary',
@@ -73,9 +70,9 @@ $container = get_theme_mod( 'understrap_container_type' );
 					)
 				); ?>
 			<?php if ( 'container' == $container ) : ?>
-			</div><!-- .container -->
+			</div><!-- /.container -->
 			<?php endif; ?>
 
-		</nav><!-- .site-navigation -->
+		</nav><!-- /.site-navigation -->
 
-	</div><!-- .wrapper-navbar end -->
+	</div><!-- /.wrapper-navbar end -->